chore: Adapt tests to OC_API refactoring
