+++ tests/lib/AppFramework/OCS/BaseResponseTest.php
@@ -7,7 +7,7 @@ declare(strict_types=1);
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-namespace Test\AppFramework\Middleware;
+namespace Test\AppFramework\OCS;
 
 use OC\AppFramework\OCS\BaseResponse;
